Update prueba_pptx and rename utils_optim
Update prueba_pptx and rename utils_optim
